Fixed bug where leaf group was not displaying
